Refactor two delete mapping methods into one, in AnimeController.java

+++ server/src/main/java/com/anime_list/anime_watch_list/controllers/AnimeController.java
@@ -75,7 +75,8 @@ public class AnimeController{
     // Delete
     // If we want to delete an anime we must remove all watch lists containing that anime FIRST,
     // as a watch list cannot have a null anime
-    @DeleteMapping(value = "animeInWatchList/{id}")
+    // Working, but error thrown if anime is not in a watchlist -> anime is still deleted though
+    @DeleteMapping(value = "/{id}")
     public ResponseEntity<String> removeAnimeInWatchListById(@PathVariable("id") Long id) {
         List<WatchList> AllWatchLists = watchListRepository.findAll();
         Optional<Anime> anime = animeRepository.findById(id);
@@ -85,17 +86,19 @@ public class AnimeController{
             if(AllWatchLists.get(i).getAnime().getName().equals(anime.get().getName())) {
                 watchListRepository.deleteById(AllWatchLists.get(i).getId());
                 animeRepository.deleteById(id);
+            } else if(AllWatchLists.size() < anime.get().getId()) {
+                animeRepository.deleteById(id);
             }
         }
 
-        return new ResponseEntity("Anime Deleted id: " + id   , HttpStatus.NOT_FOUND);
+        return new ResponseEntity("Anime Deleted " + id   , HttpStatus.NOT_FOUND);
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<String> removeAnimeNotInWatchListById(@PathVariable Long id) {
-        var found = animeRepository.findById(id);
-        animeRepository.deleteById(id);
-        return new ResponseEntity("Anime Deleted id: " + id, HttpStatus.NOT_FOUND);
-    }
+//    @DeleteMapping("/{id}")
+//    public ResponseEntity<String> removeAnimeNotInWatchListById(@PathVariable Long id) {
+//        var found = animeRepository.findById(id);
+//        animeRepository.deleteById(id);
+//        return new ResponseEntity("Anime Deleted id: " + id, HttpStatus.NOT_FOUND);
+//    }
 
 }
